Move component to old-component
