+++ src/old-component/bar-meter/index.js
