+++ src/old-component/bar-meter/item.js
