+++ src/old-component/github-icon.js
