+++ src/old-component/loading-indicator.js
