+++ src/old-component/pure-render.js
