+++ src/old-component/resizable-content.js
