+++ src/old-component/slider/grippy.js
