+++ src/old-component/slider/index.js
